Extract SingletonRecord to its own file
