+++ chia/wallet/wallet_state_manager.py
@@ -18,8 +18,9 @@ from chia_rs import AugSchemeMPL, G1Element, G2Element, PrivateKey
 from chia.consensus.block_rewards import calculate_base_farmer_reward, calculate_pool_reward
 from chia.consensus.coinbase import farmer_parent_id, pool_parent_id
 from chia.consensus.constants import ConsensusConstants
-from chia.data_layer.data_layer_wallet import DataLayerWallet, SingletonRecord
+from chia.data_layer.data_layer_wallet import DataLayerWallet
 from chia.data_layer.dl_wallet_store import DataLayerStore
+from chia.data_layer.singleton_record import SingletonRecord
 from chia.pools.pool_puzzles import (
     SINGLETON_LAUNCHER_HASH,
     get_most_recent_singleton_coin_from_coin_spend,
